# Acervo/backend/src/utils/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Controle de Acervo - Racnegê",
    "version": "1.0.0"
  },
  "paths": {
    "/users/": {
      "get": {
        "tags": ["Usuários"],
        "summary": "Obter dados do usuário",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Exemplares, exemplares disponíveis e lista de empréstimos do usuário.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserDataResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/books/": {
      "get": {
        "tags": ["Livros"],
        "summary": "Obter livros",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de livros.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetBooksInput" }
              }
            }
          }
        }
      }
    },
...
    "/loans/": {
      "get": {
        "tags": ["Empréstimos"],
        "summary": "Obter empréstimos do usuário",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de empréstimos.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/GetUserLoansResponse" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "GetUserDataResponse": {
        "type": "object",
        "properties": {
          "copies": { "type": "number", "example": 10 },
          "copiesAvailable": { "type": "number", "example": 10 },
            "loans": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "string", "example": "0167bb68-f76d-41a2-b57d-52e2622665fd" },
                "userName": { "type": "string", "example": "Nome do Usuário" },
                "bookTitle": { "type": "string", "example": "Título do Livro" },
                "ISBN": { "type": "string", "example": "978-92-95055-02-5" },
                "expirationDate": { "type": "string", "example": "2024-12-12T12:12:12.176Z" },
                "loanDate": { "type": "string", "example": "2024-12-12T12:12:12.176Z" },
                "returnDate": { "type": "string", "example": "2024-12-12T12:12:12.176Z" },
                "status": { "type": "string", "enum": ["REQUESTED", "ONGOING"], "example": "REQUESTED" },
                "postponed": { "type": "boolean", "example": "false" }
              }
            }
          }
        }
      },
...
      "GetBooksInput": {
        "type": "object",
        "properties": {
          "id": { "type": "number", "example": 1 },
          "title": { "type": "string", "example": "Título do Livro" },
          "imageUrl": { "type": "string", "example": "book-cover.jpg" },
          "category": { "type": "string", "example": "Categoria" }
        }
      },
...
      "GetUserLoansResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "number", "example": 1 },
          "userName": { "type": "string", "example": "Nome do Usuário" },
          "userId": { "type": "string", "example": "0167bb68-f76d-41a2-b57d-52e2622665fd" },
          "bookTitle": { "type": "string", "example": "Título do Livro" },
          "bookId": { "type": "number", "example": 1 },
          "ISBN": { "type": "string", "example": "978-92-95055-02-5" },
          "copyId": { "type": "number", "example": 1 },
          "expirationDate": { "type": "string", "example": "2024-12-12T12:12:12.176Z" },
          "loanDate": { "type": "string", "example": "2024-12-12T12:12:12.176Z" },
          "returnDate": { "type": "string", "example": "2024-12-12T12:12:12.176Z" },
          "status": { "type": "string", "enum": ["REQUESTED", "ONGOING"], "example": "REQUESTED" },
          "postponed": { "type": "boolean", "example": "false" },
          "archived": { "type": "boolean", "example": "false" }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}